Resolves #14, Unable to save a password on headless Linux (#15)
* 🐛 Fix issue #14 headless linux
* 🐛 Fix: check if GUI available on Linux

On Linux, gut uses gnome-keyring to save passwords. To unlock the store, gnome-keyring opens a pop-up. However, this pop-up cannot be shown when using a headless Linux server.

To solve this issue, I had to rewrite the logic to save, delete, and retrieve passwords.

If a GUI is available on Linux, gut will continue to work with gnome-keyring. However, if a GUI is not available, it will work with pass through 99designs/keyring. It is important to note that using pass requires some setup, and therefore gut will guide the user through this process.

On other platforms, gut will continue to use the zalando/go-keyring library.