Fix #16 - Improving performance of profiles loading (#18)
Changes in profiles loading
Rather than loading all profiles on each launch, gut will load them only if the command requires it.
For example, status doesn't need to load any profiles. Before this pull request, gut would load all profiles even if it's not needed.
To stop this behaviour, the ˋinit` function in the ˋprofileˋ package has been renamed.
This new function is now called on top of every exported functions of the package so that profiles are loaded when needed.

Commits:
* 🩹 Remove debug print in profile
On Linux, gut search for the path of Xorg to check if a GUI is available
If it returns an error, this means no GUI is available.
But this error must not be printed. It's only used to find out if a GUI is available

* ⚡️ Remove loading of profiles in every command
There is a lot of commands that doesn't require to load profiles
But because of how "init()" works, gut was loading every profiles and password before running the command.
This led to performance issues on Linux