✨ Add new command squash (#34)
* ✨ Add new command squash

* ✨ Add "squash" to explain command

* ♿️ Add confirm before squash

The squash command allows you to squash multiple commits into a single commit.

How it works ?

First, squash checks which commits are in the remote repository to avoid rewriting the history of others.
Then, it softly resets to the specified commit. This allows squash to leave the working directory in its current state, move the HEAD to the commit and mark the files as changed.
Finally, it amends the commit with a new message.

This results in a new commit with the content of the working directory, replacing all the commits between the HEAD and the commit specified.