✨ Add flag -v for gut version To retrieve gut version, now you can run `gut -v` or `gut version`
To be able to run `gut -v`, I had to create a controller for the root cmd.
By default, if no controller is linked to a command, cobra will show the help.

Now, root has a controller : if the -v flag is present, it fetches the version from the telemetry package and prints it.
If not, it calls cmd.help() to show the help as before

I have also created a version cmd that does the same a `gut -v`.

The version is added to the build by goreleaser thanks to ldflags.
If gut is built from source, goreleaser is not used so the gut version can't be shown.
To remediate this, gut -v falls back to "You are running a built from source version of Gut"