✨Add new capabilites to switch cmd (Fix #47)
* 🚸 Add keep changes when switching branches
Before, when the working tree wasn't clean, switching branch discarded changes
To avoid this behaviour, gut shows two options :
- Discard changes as before
- Keep the changes

To do the later, gut stash changes, switch the branch and stash pop so that the user can commit modified files

* ✨ Add switch to commit with gut switch
To get closer to Git switch behaviour, `gut switch` can now switch to a commit
First, it checks if the arg is a branch. If not, it checks if it's a commit. Finally, if it's not a commit, gut prompts to create a new branch called from the arg

Also, I've fixed a bug where gut was switching to the current branch. It's because there wasn't any check to verify the user wasn't trying to checkout the same branch.