🐛 Fix out of bound squash commit (fix #58)
To determine which commits can be squashed,
gut looks for the index of the last commit pushed.
If no commits have been pushed so far, it returns -1.

Therefore, the user can choose any commits without rewriting the history
However due to the -1 not being handled, the array was sliced by [1:-1],
rather than [1:]