🗑️ Create deprecation notice for the undo command
As pointed out by an HN commenter, the 'undo' command is not well named.
Its role is to revert the working tree to the state of the last commit,
not to undo the last commit. This is a subtle but important difference.
Therefore, we should rename the command to something more appropriate.
I think 'reset' is a good candidate.