✨ Add gut attributes cmd
As kindly requested by @FlawlessCasual17 in the #65 issue, this commit adds the gut attributes command to the CLI.
It allows users to download a .gitattributes template file from the https://github.com/gitattributes/gitattributes repository.

The gitattributes templates are embedded in the CLI for offline use.
To update the templates before a release, run bash scripts/getAttributes.sh >> src/controller/attributes.go,
and make sure to delete the old map before running the script.